+++ spring-boot/src/main/java/com/opton/spring_boot/plan/dto/Plan.java
@@ -1,4 +1,4 @@
-package com.opton.spring_boot.dto;
+package com.opton.spring_boot.plan.dto;
 
 import lombok.Data;
 import lombok.Getter;
